+++ dev/tests/drag-framer-page.tsx
@@ -1,4 +1,4 @@
-import { motion, useMotionValue } from "@framer"
+import { motion, useMotionValue, AnimateSharedLayout } from "@framer"
 import * as React from "react"
 
 /**
@@ -11,21 +11,47 @@ export const App = () => {
     const y = useMotionValue(0)
     const dummyX = useMotionValue(0)
     const dummyY = useMotionValue(0)
+    const containerRef = React.useRef<HTMLDivElement>(null)
+
+    const [count, setCount] = React.useState(0)
+
+    React.useLayoutEffect(() => {
+        const setScrollY = (yValue: number) => {
+            const element = containerRef.current
+            if (!(element instanceof HTMLDivElement)) return
+
+            element.scrollTop = -yValue
+        }
+
+        const currentY = y.get()
+        if (currentY !== 0) setScrollY(currentY)
+
+        return dummyY.onChange(setScrollY)
+    }, [dummyY])
 
     return (
-        <motion.div style={scrollContainer} layout>
+        <motion.div style={scrollContainer} layout id="root" ref={containerRef}>
             <motion.div
                 id="parent"
-                style={b}
+                style={{ ...b }}
                 drag="y"
+                layout
+                dragConstraints={containerRef}
                 _dragX={dummyX}
                 _dragY={dummyY}
+                _applyTransforms
             >
-                <motion.div style={{ ...pageContainer, x, y }} layout id="Page">
+                <motion.div
+                    style={{ ...pageContainer, x, y }}
+                    layout
+                    id="Page"
+                    onPointerEnter={() => setCount(count + 1)}
+                    _applyTransforms
+                >
                     <Page x={x} y={y} id="a" />
                     <Page x={x} y={y} id="b" />
-                    <Page x={x} y={y} />
-                    <Page x={x} y={y} />
+                    <Page x={x} y={y} id="c" />
+                    <Page x={x} y={y} id="d" />
                 </motion.div>
             </motion.div>
         </motion.div>
@@ -48,8 +74,55 @@ function Page({ x, y, id }: any) {
                 flex: "0 0 180px",
             }}
         >
-            <motion.div layout style={c} />
+            <AnimateSharedLayout>
+                <motion.div layout style={c} id={`inner-square-${id}`}>
+                    <Square id={id} />
+                </motion.div>
+            </AnimateSharedLayout>
         </motion.div>
+    )
+}
+
+function Square({ id }) {
+    const [count, setCount] = React.useState(0)
+
+    return (
+        <>
+            <motion.div
+                id={id + "-square"}
+                layout
+                style={{
+                    width: 50,
+                    height: 50,
+                    background: "#09f",
+                    borderRadius: 10,
+                }}
+                transition={{ duration: 2 }}
+                drag
+                onClick={(e) => {
+                    // e.stopPropagation()
+                    setCount(count + 1)
+                }}
+                onHoverStart={() => setCount(count + 1)}
+            />
+            <motion.div
+                id={id + "-square-green"}
+                layout
+                style={{
+                    width: 50,
+                    height: 50,
+                    background: "green",
+                    borderRadius: 10,
+                }}
+                transition={{ duration: 2 }}
+                drag
+                onClick={(e) => {
+                    // e.stopPropagation()
+                    setCount(count + 1)
+                }}
+                onHoverStart={() => setCount(count + 1)}
+            />
+        </>
     )
 }
 
@@ -57,7 +130,7 @@ const scrollContainer: React.CSSProperties = {
     position: "absolute",
     top: 100,
     left: 100,
-    width: 200,
+    width: 600,
     height: 500,
     overflow: "hidden",
 }
@@ -81,7 +154,7 @@ const b = {
     top: 100,
     left: 100,
     width: "100%",
-    height: 500,
+    height: 1000,
     borderRadius: 10,
 }