+++ src/components/AnimateSharedLayout/utils/stack.ts
@@ -1,6 +1,6 @@
 import { Presence, SharedLayoutAnimationConfig } from "../types"
-import { AxisBox2D, Point2D } from "../../../types/geometry"
-import { ResolvedValues, VisualElement } from "../../../render/types"
+import { Point2D } from "../../../types/geometry"
+import { ResolvedValues, Snapshot, VisualElement } from "../../../render/types"
 import { elementDragControls } from "../../../gestures/drag/VisualElementDragControls"
 import { createCrossfader } from "./crossfader"
 
@@ -15,7 +15,11 @@ export interface LayoutStack {
     getLead(): VisualElement | undefined
     updateSnapshot(): void
     clearSnapshot(): void
-    animate(element: VisualElement, crossfade: boolean): void
+    animate(
+        element: VisualElement,
+        crossfade: boolean,
+        isControlled?: boolean
+    ): void
     updateLeadAndFollow(): void
 }
 
@@ -31,13 +35,15 @@ export function layoutStack(): LayoutStack {
     let prevState: StackState = { ...state }
 
     let prevValues: ResolvedValues | undefined
-    let prevViewportBox: AxisBox2D | undefined
+    let snapshot: Snapshot | undefined
     let prevDragCursor: Point2D | undefined
     const crossfader = createCrossfader()
     let needsCrossfadeAnimation = false
 
     function getFollowViewportBox() {
-        return state.follow ? state.follow.prevViewportBox : prevViewportBox
+        return state.follow
+            ? state.follow.snapshot?.viewportBox
+            : snapshot?.viewportBox
     }
 
     function getFollowLayout() {
@@ -66,7 +72,7 @@ export function layoutStack(): LayoutStack {
             prevValues = crossfader.isActive()
                 ? crossfader.getLatestValues()
                 : state.lead.getLatestValues()
-            prevViewportBox = state.lead.prevViewportBox
+            snapshot = state.lead.snapshot
 
             const dragControls = elementDragControls.get(state.lead)
             if (dragControls && dragControls.isDragging) {
@@ -74,7 +80,7 @@ export function layoutStack(): LayoutStack {
             }
         },
         clearSnapshot() {
-            prevDragCursor = prevViewportBox = undefined
+            prevDragCursor = snapshot = undefined
         },
         updateLeadAndFollow() {
             prevState = { ...state }
@@ -141,9 +147,11 @@ export function layoutStack(): LayoutStack {
                     needsCrossfadeAnimation = false
                     const transition = child.getDefaultTransition()
 
-                    child.presence === Presence.Entering
-                        ? crossfader.toLead(transition)
-                        : crossfader.fromLead(transition)
+                    if (child.presence === Presence.Entering) {
+                        crossfader.toLead(transition)
+                    } else if (child.presence === Presence.Exiting) {
+                        crossfader.fromLead(transition)
+                    }
                 }
 
                 child.notifyLayoutReady(config)