+++ src/utils/geometry/delta-apply.ts
@@ -243,7 +243,7 @@ export function applyTreeDeltas(
         applyBoxDelta(box, delta)
 
         // If this is a draggable ancestor, also incorporate the node's transform to the layout box
-        if (isDraggable(node)) {
+        if (isDraggable(node) || node.getProps()._applyTransforms) {
             applyBoxTransforms(box, box, node.getLatestValues())
         }
     }