+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Framer Motion adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.1.18] Unreleased
+
+### Fixed
+
+-   Various layout animation fixes.
+-   Explicit handling of NaN values in layout projection.
+-   Fixed an issue where MotionConfig would not update when the provided transition was updated.
+
 ## [4.1.17] 2021-05-17
 
 ### Fixed